Develop

+++ src/pages/AddContact/AdminContact.js
@@ -11,6 +11,7 @@ import ServiceApi from "../../services/Service";
 import { useDispatch } from "react-redux";
 import { fetchContact } from "../../action";
 import AddContact from "./AddContact";
+import { getCookies } from "../../utils/Utility";
 
 const { confirm } = Modal;
 
@@ -26,6 +27,8 @@ const AdminContacts = function ({ currentLang }) {
 
   const { t } = useTranslation();
   const dispatch = useDispatch();
+  const checkAdmin = getCookies("user_token")?.user?.roles?.find(item=>item.calendarId==="CULTURE_OUTAOUAIS")
+
 
   const eventTableHeader = [
     {
@@ -71,11 +74,13 @@ const AdminContacts = function ({ currentLang }) {
       key: "hasDependency",
       width:100,
       render: (e, record) => (
+        checkAdmin && (checkAdmin.role === "ADMIN" || checkAdmin.role === "SUPER_ADMIN")?
         <DeleteOutlined
           style={{fontSize:"23px"}}
           onClick={(event) => handleDelete(record, event)}
          
         />
+        :<></>
       ),
     },
   ];
@@ -218,6 +223,8 @@ const AdminContacts = function ({ currentLang }) {
                 return {
                   onClick: (event) => {
                     event.stopPropagation()
+                    if(getCookies("user_token")?.user?.id===record.creator?.userId ||(checkAdmin && (checkAdmin.role === "EDITOR" || checkAdmin.role === "ADMIN" || checkAdmin.role === "SUPER_ADMIN")))
+
                     navigate(`/admin/add-contact/?id=${record.uuid}`);
                     
                   },
